Refactoring Code
`
1. Added services class to encapsulate logic
2. Moved calendar function to dedicated utils folder
`
